+++ example/lib/main.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+import 'streambuilder_test.dart';
 import 'package:geoflutterfire/geoflutterfire.dart';
 import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:rxdart/rxdart.dart';
@@ -30,7 +31,7 @@ class _MyAppState extends State<MyApp> {
   // firestore init
   final _firestore = FirebaseFirestore.instance;
   late Geoflutterfire geo;
-  late Stream<List<DocumentSnapshot<Map<String, dynamic>>>> stream;
+  late Stream<List<DocumentSnapshot>> stream;
   var radius = BehaviorSubject.seeded(100.0);
 
   //Map markers
@@ -44,9 +45,7 @@ class _MyAppState extends State<MyApp> {
     stream = radius.switchMap((rad) {
       var collectionReference = _firestore.collection('locations');
       //          .where('name', isEqualTo: 'darshan');
-      return geo
-          .collection(collectionRef: collectionReference)
-          .within(center: center, radius: rad, field: 'position');
+      return geo.collection(collectionRef: collectionReference).within(center: center, radius: rad, field: 'position');
 
       /** Example to specify nested object
           var collectionReference = _firestore.collection('nestedLocations');
@@ -165,7 +164,21 @@ class _MyAppState extends State<MyApp> {
                     double lng = double.parse(_longitudeController.text);
                     _addNestedPoint(lat, lng);
                   },
+                ),
+                MaterialButton(
+                  color: Colors.blueGrey,
+                  child: Text(
+                    'Stream Test ',
+                    style: TextStyle(color: Colors.white),
+                  ),
+                  onPressed: () {
+                    Navigator.push(
+                      context,
+                      MaterialPageRoute(builder: (context) => StreamTestWidget()),
+                    );
+                  },
                 )
+
               ],
             ),
           ),
@@ -176,7 +189,7 @@ class _MyAppState extends State<MyApp> {
 
   void _onMapCreated(GoogleMapController controller) {
     _mapController.complete(controller);
-    stream.listen((List<DocumentSnapshot<Map<String, dynamic>>> documentList) {
+    stream.listen((List<DocumentSnapshot> documentList) {
       _updateMarkers(documentList);
     });
   }
@@ -193,12 +206,12 @@ class _MyAppState extends State<MyApp> {
 
   void _addPoint(double lat, double lng) {
     GeoFirePoint geoFirePoint = geo.point(latitude: lat, longitude: lng);
-    _firestore
-        .collection('locations')
-        .add({'name': 'random name', 'position': geoFirePoint.data}).then((_) {
+    _firestore.collection('locations').add({'name': 'random name', 'position': geoFirePoint.data}).then((_) {
       print('added ${geoFirePoint.hash} successfully');
     });
-    setState(() {});
+    setState(() {
+
+    });
   }
 
   //example to add geoFirePoint inside nested object
@@ -225,9 +238,8 @@ class _MyAppState extends State<MyApp> {
     });
   }
 
-  void _updateMarkers(
-      List<DocumentSnapshot<Map<String, dynamic>>> documentList) {
-    documentList.forEach((DocumentSnapshot<Map<String, dynamic>> document) {
+  void _updateMarkers(List<DocumentSnapshot> documentList) {
+    documentList.forEach((DocumentSnapshot document) {
       GeoPoint point = document['position']['geopoint'];
       _addMarker(point.latitude, point.longitude);
     });